requirements.txt
Flask
torch
torchvision
Pillow
numpy
scikit-learn
matplotlib
seaborn
waitress